requirements.txt
flask
openai
openai-whisper
torch
numpy
soundfile
sqlalchemy
python-dotenv